+++ packages/ui5-test-writer/templates/freestyle/webapp/test/1.71.0/unit/unitTests.qunit.ts
@@ -1,0 +1,10 @@
+/* global QUnit */
+// https://api.qunitjs.com/config/autostart/
+QUnit.config.autostart = false;
+
+// import all your QUnit tests here
+void Promise.all([
+import("unit/controller/<%= viewName %>Page.controller")
+]).then(() => {
+	QUnit.start();
+});